// apps/api/src/modules/openai/dto/completions-dtos/ChoiceDto.ts
import { ApiExtraModels, ApiProperty, getSchemaPath } from '@nestjs/swagger';
import { IsIn, IsNumber, ValidateNested } from 'class-validator';
import { Type } from 'class-transformer';

import { LogprobsDto } from './LogprobsDto';
import { ChatCompletionMessageDto } from './ChatCompletionMessageDto';

@ApiExtraModels(
  LogprobsDto,
)
export class ChoiceDto {
  /**
   * The reason the model stopped generating tokens. This will be `stop` if the model
   * hit a natural stop point or a provided stop sequence, `length` if the maximum
   * number of tokens specified in the request was reached, `content_filter` if
   * content was omitted due to a flag from our content filters, `tool_calls` if the
   * model called a tool, or `function_call` (deprecated) if the model called a
   * function.
   */
  @ApiProperty({
    description: `The reason the model stopped generating tokens. This will be \`stop\` if the model
  hit a natural stop point or a provided stop sequence, \`length\` if the maximum
  number of tokens specified in the request was reached, \`content_filter\` if
  content was omitted due to a flag from our content filters, \`tool_calls\` if the
  model called a tool, or \`function_call\` (deprecated) if the model called a
  function.`,
    enum: ['stop', 'length', 'tool_calls', 'content_filter', 'function_call'],
  })
  @IsIn(['stop', 'length', 'tool_calls', 'content_filter', 'function_call'])
  finish_reason!: 'stop' | 'length' | 'tool_calls' | 'content_filter' | 'function_call';

  /** The index of the choice in the list of choices. */
  @ApiProperty({
    description: `The index of the choice in the list of choices.`,
    type: 'number',
  })
  @IsNumber()
  index!: number;

  /** Log probability information for the choice. */
  @ApiProperty({
    description: `Log probability information for the choice.`,
    type: () => LogprobsDto,
  })
  logprobs!: null | LogprobsDto;

  /** A chat completion message generated by the model. */
  @ApiProperty({
    description: `A chat completion message generated by the model.`,
    type: () => ChatCompletionMessageDto,
  })
  @ValidateNested()
  @Type(() => ChatCompletionMessageDto)
  message!: ChatCompletionMessageDto;
}